Made the Controllers only return the data that concers you Only your adress or paiment etc ...

+++ Project Manager 2/Project Manager/Local Server API/Controllers/IssuesController.cs
@@ -6,8 +6,10 @@ using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Text;
 using System.Web.Http;
 using System.Web.Http.Description;
+using System.Web.UI;
 using DataAccess.Models;
 using Local_Server_API.Models;
 
@@ -18,9 +20,9 @@ namespace Local_Server_API.Controllers
         private Local_DB_Model db = new Local_DB_Model();
 
         [AuthorizaAttr]
-        public IQueryable<Issue> GetIssue()
+        public ICollection<Issue> GetIssue()
         {
-            return db.Issues;
+            return GetTeamIssues(ActionContext.Request.Headers.Authorization.Parameter);
         }
 
         [AuthorizaAttr]
@@ -48,6 +50,11 @@ namespace Local_Server_API.Controllers
             if (id != issue.IssueID)
             {
                 return BadRequest();
+            }
+
+            if (GetTeamIssues(ActionContext.Request.Headers.Authorization.Parameter).Where(I => I.IssueID == id).FirstOrDefault() is null)
+            {
+                return Unauthorized();
             }
 
             db.Entry(issue).State = EntityState.Modified;
@@ -80,6 +87,11 @@ namespace Local_Server_API.Controllers
                 return BadRequest(ModelState);
             }
 
+            if (issue.Task1.Project1.Team1.TeamID != GetUserFromAuthHeader(ActionContext.Request.Headers.Authorization.Parameter).Team)
+            {
+                return Unauthorized();
+            }
+
             db.Issues.Add(issue);
             db.SaveChanges();
 
@@ -94,6 +106,11 @@ namespace Local_Server_API.Controllers
             if (issue == null)
             {
                 return NotFound();
+            }
+
+            if (GetTeamIssues(ActionContext.Request.Headers.Authorization.Parameter).Where(I => I.IssueID == id).FirstOrDefault() is null)
+            {
+                return Unauthorized();
             }
 
             db.Issues.Remove(issue);
@@ -114,6 +131,34 @@ namespace Local_Server_API.Controllers
         private bool IssueExists(int id)
         {
             return db.Issues.Count(e => e.IssueID == id) > 0;
+        }
+
+        [NonAction]
+        /// <summary>
+        /// Gets the user based on the Auth header parameter
+        /// </summary>
+        /// <param name="AuthParameter"><code>ActionContext.Request.Headers.Authorization.Parameter</code></param>
+        /// <returns></returns>
+        public User GetUserFromAuthHeader(string AuthParameter)
+        {
+            string RequestingUID = Encoding.UTF8.GetString(Convert.FromBase64String(AuthParameter)).Split(':')[0];
+            return db.Users.Where(U => U.UserName == RequestingUID).FirstOrDefault();
+        }
+
+        private List<Issue> GetTeamIssues(string AuthParameter)
+        {
+            var RequestingUser = GetUserFromAuthHeader(AuthParameter);
+
+            List<Issue> Issues = new List<Issue>();
+
+            RequestingUser.Team1.Projects.ToList()
+                .ForEach(p => p.Tasks.ToList()
+                    .ForEach(t => t.Issues.ToList()
+                                .ForEach(i => Issues.Add(i))
+                            )
+                        );
+
+            return Issues;
         }
     }
 }
